extract summary data into hook
